Reorganize add_position and add_video and tests
