+++ tests/test_convert_video.py
@@ -1,8 +1,85 @@
 from datetime import datetime
 from dateutil import tz
+from zoneinfo import ZoneInfo
 from pynwb import NWBFile
+from pathlib import Path
+from ndx_franklab_novela import CameraDevice
+from hdmf.common.table import DynamicTable, VectorData
 
-from jdb_to_nwb.convert_video import add_video
+from jdb_to_nwb.convert_video import add_video, add_hex_centroids, add_camera
+
+
+def test_add_hex_centroids(dummy_logger):
+    """
+    Test the add_hex_centroids function
+    """
+
+    test_data_dir = Path("tests/test_data/downloaded/IM-1478/07252022")
+
+    metadata = {}
+    metadata["video"] = {}
+    metadata["video"]["pixels_per_cm"] = 3.14
+    metadata["video"]["hex_centroids_file_path"] = test_data_dir / "hex_coordinates_IM-1478_07252022.csv"
+
+    nwbfile = NWBFile(
+        session_description="Mock session",
+        session_start_time=datetime.now(tz.tzlocal()),
+        identifier="mock_session",
+    )
+
+    # Add centroids table to the nwb
+    add_hex_centroids(nwbfile=nwbfile, metadata=metadata, logger=dummy_logger)
+
+    # Check that behavior processing module has been added
+    assert "behavior" in nwbfile.processing
+
+    # Check that the "hex_centroids" table exists in the behavior processing module
+    behavior_module = nwbfile.processing["behavior"]
+    assert "hex_centroids" in behavior_module.data_interfaces
+    hex_centroids_table = behavior_module.data_interfaces["hex_centroids"]
+
+    assert isinstance(hex_centroids_table, DynamicTable)
+    assert hex_centroids_table.name == "hex_centroids"
+    assert hex_centroids_table.description == "Centroids of each hex in the maze (in video pixel coordinates)"
+
+    # The table should have data for all 49 hex centroids
+    assert len(hex_centroids_table) == 49
+
+    # Check that the table contains the correct columns
+    for column in hex_centroids_table.columns:
+        assert isinstance(column, VectorData)
+    expected_columns = {"hex", "x", "y", "x_meters", "y_meters"}
+    assert set(hex_centroids_table.colnames) == expected_columns, (
+        f"Hex centroid columns {set(hex_centroids_table.colnames)} "
+        f"did not match expected {expected_columns}"
+    )
+
+
+def test_add_camera(dummy_logger):
+    """
+    Test the add_camera function
+    """
+
+    metadata = {}
+    metadata["video"] = {}
+    metadata["video"]["pixels_per_cm"] = 3.14
+
+    nwbfile = NWBFile(
+        session_description="Mock session",
+        session_start_time=datetime.now(tz.tzlocal()),
+        identifier="mock_session",
+    )
+
+    # Add a CameraDevice to the nwb
+    add_camera(nwbfile=nwbfile, metadata=metadata)
+
+    # Check that the CameraDevice has been added
+    assert "camera_device 1" in nwbfile.devices
+    camera_device = nwbfile.devices["camera_device 1"]
+    assert isinstance(camera_device, CameraDevice)
+    assert camera_device.camera_name == "maze_camera"
+    assert camera_device.meters_per_pixel == metadata["video"]["pixels_per_cm"]
+    assert camera_device.manufacturer == "Logitech"
 
 
 def test_add_video(dummy_logger):
@@ -12,6 +89,7 @@ def test_add_video(dummy_logger):
     metadata["video"] = {}
     metadata["video"]["video_file_path"] = "tests/test_data/downloaded/IM-1478/07252022/Behav_Vid0.avi"
     metadata["video"]["video_timestamps_file_path"] = "tests/test_data/downloaded/IM-1478/07252022/testvidtimes0.csv"
+    metadata["datetime"] = datetime.strptime("07252022", "%m%d%Y").replace(tzinfo=ZoneInfo("America/Los_Angeles")) 
 
     test_output_video_path = "tests/test_data/downloaded/output/test_video.mp4"
 
@@ -74,6 +152,7 @@ def test_add_video_with_incomplete_metadata(capsys, dummy_logger):
 
     # Create a test metadata dictionary with a video field but no video data
     metadata["video"] = {}
+    metadata["datetime"] = datetime.strptime("07252022", "%m%d%Y").replace(tzinfo=ZoneInfo("America/Los_Angeles")) 
 
     # Call the add_video function with 'video' key in metadata, but no video subfields
     add_video(nwbfile=nwbfile, metadata=metadata, 
@@ -84,3 +163,4 @@ def test_add_video_with_incomplete_metadata(capsys, dummy_logger):
     # This should print with no error, because it is ok if we have 'video' but not these subfields, 
     # because maybe we only have DLC and not raw video
     assert "Skipping video file conversion" in captured.out
+    assert "No subfield 'hex_centroids_file_path' found in video metadata!" in captured.out
